Tests involving docfx moved to their own class
This enabled using OneTimeSetUp for running docfx once, which is much cleaner than the previous way.
