+++ Assets/Tests/InputSystem/DocumentationBasedAPIVerficationTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3cd00bf0a5aeb024cbf6d4441b08f871
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
